usm: Remove non-linux oriented files
