--- pkg/network/protocols/http2/dummy.go
@@ -1,9 +1,0 @@
-// Unless explicitly stated otherwise all files in this repository are licensed
-// under the Apache License Version 2.0.
-// This product includes software developed at Datadog (https://www.datadoghq.com/).
-// Copyright 2016-present Datadog, Inc.
-
-// On Windows/mac the package his empty, so having a dummy file to ensure the package is not invalid on Windows/mac.
-
-// Package http2 provides a HTTP/2 implementation for the network package for supporting HTTP/2 and gRPC monitoring with USM.
-package http2
