+++ pkg/network/protocols/http2/raw_traffic.go
@@ -3,7 +3,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-//go:build test
+//go:build test && linux_bpf
 
 package http2
