+++ pkg/network/protocols/postgres/debugging/debugging.go
@@ -3,6 +3,8 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
+//go:build linux_bpf
+
 // Package debugging provides debug-friendly representations of internal data structures
 package debugging
